Add interruption routine to handle external interrupt
This module is particularly special because it can be used to
detect rising and falling edge of the incoming signal.

It will be necessary to comunicate with the RX signal and controll the
DC motor speed with PWM; perhaps a better way to approach this type of
interaction would be using the SPI communication module, but I have
a feeling that this would be a bit overhead.

An 8-bit register counter [TCCR2] is used to set the duty cycle
definition.