Enable external interrupt on PD5
The modulated signal comming from the RX generates an interrupt
request on border rising and falling edge clk cycle. A timer [TCCR2] is set
to count every cycle of the external clock divided by a 256-prescaler.

When a change of state (from LOW to a HIGH) is detected, the value
of the timer2 register [TCNT2] is fetched filtered and connected to the OCR1A
register (leftmost count value) to determine the duty cycle; that way
the motor is controlled by the PB1 PWM output signal.