Copy PWM signal from channel OC1A to OC1B
The same channel (channel 1 on the radio) controls the speed in both direction,
forwards and backwards.

In order to avoid having to implement another interruption routine,
alocating a new interruption vector (OC1A/OC1B compare match
interrupt vector is a good candidate); a simple solution is to mirror
the signal in channel A to channel B of the Timer/counter1.