Fix bugs in the speed controller unit
Capturing signal by external interrupt was never a good idea. Even
though the comunication was super slow (50Hz) that took almost all CPU
time. Doing something other than the controling one motor were
unmanageble.

A fix is to implement a simple serial comunication with clock
synchronization. As the actual bytes are not very import (the important
part is the width of pulses each clock cycle) syncing is not a big deal.
Calculating the parameters can be done empirically without many losses.
Digitally filtering the signal helps a lot. A moving average was
implemented.