Simplifying a lot of stuff, comming back to external interrupts
This patch simplifies a lot of the parameters introduced by b467cff
Most of these parameters did not survive a reset. The parameters is some
sort of "fine tune", so the fine tuning should be performed every reset,
this is tottaly impractical and I just figured out after spending a few hours
tweaking it.

So turns ou that syncing by external interrupts isn't that bad if done
right, and if the frequency of external clock isn't too great;